app/requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
gunicorn==20.0.4
psycopg2-binary==2.8.6
